acra_webhook/requirements.txt
requests
sqlalchemy
apscheduler
uvicorn
fastapi
python-dotenv
psycopg2-binary
pydantic
pydantic-settings

pytest
